id 타입 Integer에서 Long으로 변경

+++ src/main/java/com/allrounders/goalkeeper/domain/OXGame.java
@@ -9,7 +9,7 @@ import jakarta.persistence.Id;
 public class OXGame {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer oxId;
+    private Long oxId;
 
     @Column(nullable = false)
     private String question;
